[Web] Renames Public area in Publics
