--- src/CriThink.Server/CriThink.Server.Web/Areas/Public/Views/_ViewImports.cshtml
@@ -1,3 +1,0 @@
-@using CriThink.Server.Web.Areas.Public.ViewModel.Identity
-@using CriThink.Server.Web.Areas.Public.ViewModel.Shared
-@addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
