+++ src/CriThink.Server/CriThink.Server.Web/Areas/Publics/Controllers/IdentityController.cs
@@ -3,13 +3,13 @@ using System.Threading.Tasks;
 using CriThink.Common.Endpoints;
 using CriThink.Common.Helpers;
 using CriThink.Server.Core.Interfaces;
-using CriThink.Server.Web.Areas.Public.ViewModel.Identity;
-using CriThink.Server.Web.Areas.Public.ViewModel.Shared;
+using CriThink.Server.Web.Areas.Publics.ViewModel;
+using CriThink.Server.Web.Areas.Publics.ViewModel.Identity;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
 
-namespace CriThink.Server.Web.Areas.Public.Controllers
+namespace CriThink.Server.Web.Areas.Publics.Controllers
 {
     [Area("Public")]
     [ApiExplorerSettings(IgnoreApi = true)]
