+++ src/CriThink.Server/CriThink.Server.Web/Areas/Publics/ViewModel/Identity/ResetPasswordViewModel.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
 
-namespace CriThink.Server.Web.Areas.Public.ViewModel.Identity
+namespace CriThink.Server.Web.Areas.Publics.ViewModel.Identity
 {
     public class ResetPasswordViewModel : IValidatableObject
     {
