+++ src/CriThink.Server/CriThink.Server.Web/Areas/Publics/Views/Identity/ResetPassword.cshtml
@@ -1,4 +1,4 @@
-﻿@model ResetPasswordViewModel
+﻿@model CriThink.Server.Web.Areas.Publics.ViewModel.Identity.ResetPasswordViewModel
 @{
     Layout = "_Layout";
     ViewData["Title"] = "Change password";
