+++ src/CriThink.Server/CriThink.Server.Web/Areas/Publics/Views/Shared/Message.cshtml
@@ -1,4 +1,4 @@
-﻿@model MessageViewModel
+﻿@model CriThink.Server.Web.Areas.Publics.ViewModel.MessageViewModel
 @{
     Layout = "_Layout";
     ViewData["Title"] = @Model.Title;
