+++ src/CriThink.Server/CriThink.Server.Web/Areas/Publics/Views/_ViewImports.cshtml
@@ -1,0 +1,1 @@
+@addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
