+++ src/CriThink.Server/CriThink.Server.Web/Facades/DebunkingNewsServiceFacade.cs
@@ -2,8 +2,8 @@
 using System.Threading.Tasks;
 using CriThink.Common.Endpoints.DTOs.Admin;
 using CriThink.Server.Core.Interfaces;
+using CriThink.Server.Web.Areas.BackOffice.ViewModels;
 using CriThink.Server.Web.Areas.BackOffice.ViewModels.DebunkingNews;
-using CriThink.Server.Web.Areas.BackOffice.ViewModels.Shared;
 
 namespace CriThink.Server.Web.Facades
 {
